+++ torchao/quantization/pt2e/__init__.py
@@ -70,7 +70,6 @@ from .observer import (
     TorchAODType,
     UniformQuantizationObserverBase,
     ZeroPointDomain,
-    get_block_size,
 )
 
 for _f in [
@@ -149,7 +148,6 @@ __all__ = [
     "PerToken",
     "TorchAODType",
     "ZeroPointDomain",
-    "get_block_size",
     "default_fake_quant",
     "default_dynamic_fake_quant",
 ]
